chore(flutter): setup example app
