+++ flutter/measure_flutter/android/build.gradle
@@ -1,15 +1,15 @@
-group = "com.example.measure_flutter"
-version = "1.0-SNAPSHOT"
+group = "sh.measure.flutter"
+version = "0.1.0-SNAPSHOT"
 
 buildscript {
-    ext.kotlin_version = "1.8.22"
+    ext.kotlin_version = "1.9.21"
     repositories {
         google()
         mavenCentral()
     }
 
     dependencies {
-        classpath("com.android.tools.build:gradle:8.1.0")
+        classpath("com.android.tools.build:gradle:8.7.3")
         classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version")
     }
 }
@@ -25,7 +25,7 @@ apply plugin: "com.android.library"
 apply plugin: "kotlin-android"
 
 android {
-    namespace = "com.example.measure_flutter"
+    namespace = "sh.measure.flutter"
 
     compileSdk = 35
