+++ flutter/measure_flutter/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.example.measure_flutter">
+  package="sh.measure.flutter">
 </manifest>
