+++ flutter/measure_flutter/example/android/build.gradle
@@ -2,6 +2,7 @@ allprojects {
     repositories {
         google()
         mavenCentral()
+        mavenLocal()
     }
 }
